+++ tests/data_io/implementations/test_numpy_io.py
@@ -24,13 +24,14 @@ __maintainer__ = "Malte Storm"
 __status__ = "Production"
 
 
-import os
 import shutil
 import tempfile
 import unittest
+from pathlib import Path
 
 import numpy as np
 
+from pydidas.core import FileReadError
 from pydidas.core.constants import NUMPY_EXTENSIONS
 from pydidas.data_io.implementations.numpy_io import NumpyIo
 
@@ -38,8 +39,8 @@ from pydidas.data_io.implementations.numpy_io import NumpyIo
 class TestNumpyIo(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls._path = tempfile.mkdtemp()
-        cls._fname = os.path.join(cls._path, "test.npy")
+        cls._path = Path(tempfile.mkdtemp())
+        cls._fname = cls._path.joinpath("test.npy")
         cls._data_shape = (12, 13, 14, 15)
         cls._data = np.random.random(cls._data_shape)
         np.save(cls._fname, cls._data)
@@ -63,19 +64,30 @@ class TestNumpyIo(unittest.TestCase):
         _data = NumpyIo.import_from_file(self._fname)
         self.assertTrue(np.allclose(_data, self._data))
 
+    def test_import_from_file__wrong_name(self):
+        with self.assertRaises(FileReadError):
+            NumpyIo.import_from_file(self._fname.joinpath("dummy"), datatype=np.float64)
+
+    def test_import_from_file__wrong_type(self):
+        _fname_new = self._path.joinpath("test2.dat")
+        with open(_fname_new, "w") as f:
+            f.write("now it's just an ASCII text file.")
+        with self.assertRaises(FileReadError):
+            NumpyIo.import_from_file(_fname_new, datatype=np.float64)
+
     def test_export_to_file__file_exists(self):
         with self.assertRaises(FileExistsError):
             NumpyIo.export_to_file(self._fname, self._data)
 
     def test_export_to_file__file_exists_and_overwrite(self):
-        _fname = os.path.join(self._path, "test_new.npy")
+        _fname = self._path.joinpath("test_new.npy")
         NumpyIo.export_to_file(_fname, self._data)
         NumpyIo.export_to_file(_fname, self._data[:11], overwrite=True)
         _data = NumpyIo.import_from_file(_fname)
         self.assertEqual(_data.shape, (11,) + self._data_shape[1:])
 
     def test_export_to_file__simple(self):
-        _fname = os.path.join(self._path, "test_fname.npy")
+        _fname = self._path.joinpath("test_fname.npy")
         NumpyIo.export_to_file(_fname, self._data)
         _data = NumpyIo.import_from_file(_fname)
         self.assertTrue(np.allclose(_data, self._data))