+++ tests/data_io/implementations/test_raw_io.py
@@ -24,13 +24,14 @@ __maintainer__ = "Malte Storm"
 __status__ = "Production"
 
 
-import os
 import shutil
 import tempfile
 import unittest
+from pathlib import Path
 
 import numpy as np
 
+from pydidas.core import FileReadError
 from pydidas.core.constants import BINARY_EXTENSIONS
 from pydidas.data_io.implementations.raw_io import RawIo
 
@@ -38,8 +39,8 @@ from pydidas.data_io.implementations.raw_io import RawIo
 class TestRawIo(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls._path = tempfile.mkdtemp()
-        cls._fname = os.path.join(cls._path, "test.npy")
+        cls._path = Path(tempfile.mkdtemp())
+        cls._fname = cls._path.joinpath("test.dat")
         cls._data_shape = (12, 13, 14)
         cls._data = np.random.random(cls._data_shape)
         cls._index = 0
@@ -70,19 +71,30 @@ class TestRawIo(unittest.TestCase):
         _data = RawIo.import_from_file(self._fname, **self._read_kws)
         self.assertTrue(np.allclose(_data, self._data))
 
+    def test_import_from_file__wrong_name(self):
+        with self.assertRaises(FileReadError):
+            RawIo.import_from_file(self._fname.joinpath("dummy"), datatype=np.float64)
+
+    def test_import_from_file__wrong_type(self):
+        _fname_new = self._path.joinpath("test2.dat")
+        with open(_fname_new, "w") as f:
+            f.write("now it's just an ASCII text file.")
+        with self.assertRaises(FileReadError):
+            RawIo.import_from_file(_fname_new, datatype=np.float64)
+
     def test_export_to_file__file_exists(self):
         with self.assertRaises(FileExistsError):
             RawIo.export_to_file(self._fname, self._data)
 
     def test_export_to_file__file_exists_and_overwrite(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         RawIo.export_to_file(_fname, self._data[:11])
         RawIo.export_to_file(_fname, self._data, overwrite=True)
         _data = RawIo.import_from_file(self._fname, **self._read_kws)
         self.assertEqual(_data.shape, self._data_shape)
 
     def test_export_to_file__simple(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         RawIo.export_to_file(_fname, self._data)
         _data = RawIo.import_from_file(_fname, **self._read_kws)
         self.assertTrue(np.allclose(_data, self._data))