+++ tests/data_io/implementations/test_tiff_io.py
@@ -24,14 +24,15 @@ __maintainer__ = "Malte Storm"
 __status__ = "Production"
 
 
-import os
 import shutil
 import tempfile
 import unittest
+from pathlib import Path
 
 import numpy as np
 import skimage.io
 
+from pydidas.core import FileReadError
 from pydidas.core.constants import TIFF_EXTENSIONS
 from pydidas.data_io.implementations.tiff_io import TiffIo
 
@@ -39,8 +40,8 @@ from pydidas.data_io.implementations.tiff_io import TiffIo
 class TestTiffIo(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        cls._path = tempfile.mkdtemp()
-        cls._fname = os.path.join(cls._path, "test.tif")
+        cls._path = Path(tempfile.mkdtemp())
+        cls._fname = cls._path.joinpath("test.tif")
         cls._data_shape = (12, 13)
         cls._data = np.random.random(cls._data_shape)
         cls._index = 0
@@ -70,25 +71,35 @@ class TestTiffIo(unittest.TestCase):
         _data = TiffIo.import_from_file(self._fname)
         self.assertTrue(np.allclose(_data, self._data))
 
+    def test_import_from_file__wrong_name(self):
+        with self.assertRaises(FileReadError):
+            TiffIo.import_from_file(self._fname.joinpath("dummy"))
+
+    def test_import_from_file__wrong_type(self):
+        with open(self._fname, "w") as f:
+            f.write("now it's just an ASCII text file.")
+        with self.assertRaises(FileReadError):
+            TiffIo.import_from_file(self._fname)
+
     def test_export_to_file__file_exists(self):
         with self.assertRaises(FileExistsError):
             TiffIo.export_to_file(self._fname, self._data)
 
     def test_export_to_file__file_exists_and_overwrite(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         TiffIo.export_to_file(_fname, self._data)
         TiffIo.export_to_file(_fname, self._data[:11], overwrite=True)
         _data = TiffIo.import_from_file(_fname)
         self.assertEqual(_data.shape, (11,) + self._data_shape[1:])
 
     def test_export_to_file__simple(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         TiffIo.export_to_file(_fname, self._data)
         _data = TiffIo.import_from_file(_fname)
         self.assertTrue(np.allclose(_data, self._data))
 
     def test_export_to_file__8bit_int(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         _raw = (np.random.random((11, 12, 13)) * 125).astype(np.int8)
         TiffIo.export_to_file(_fname, _raw)
         _data = TiffIo.import_from_file(_fname)
@@ -96,7 +107,7 @@ class TestTiffIo(unittest.TestCase):
         self.assertEqual(_raw.dtype, _data.dtype)
 
     def test_export_to_file__8bit_uint(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         _raw = (np.random.random((11, 12, 13)) * 125).astype(np.uint8)
         TiffIo.export_to_file(_fname, _raw)
         _data = TiffIo.import_from_file(_fname)
@@ -104,7 +115,7 @@ class TestTiffIo(unittest.TestCase):
         self.assertEqual(_raw.dtype, _data.dtype)
 
     def test_export_to_file__16bit_int(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         _raw = (np.random.random((11, 12, 13)) * 1257).astype(np.int16)
         TiffIo.export_to_file(_fname, _raw)
         _data = TiffIo.import_from_file(_fname)
@@ -112,7 +123,7 @@ class TestTiffIo(unittest.TestCase):
         self.assertEqual(_raw.dtype, _data.dtype)
 
     def test_export_to_file__16bit_uint(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         _raw = (np.random.random((11, 12, 13)) * 1257).astype(np.uint16)
         TiffIo.export_to_file(_fname, _raw)
         _data = TiffIo.import_from_file(_fname)
@@ -120,7 +131,7 @@ class TestTiffIo(unittest.TestCase):
         self.assertEqual(_raw.dtype, _data.dtype)
 
     def test_export_to_file__32bit_int(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         _raw = (np.random.random((11, 12, 13)) * 12573).astype(np.int32)
         TiffIo.export_to_file(_fname, _raw)
         _data = TiffIo.import_from_file(_fname)
@@ -128,7 +139,7 @@ class TestTiffIo(unittest.TestCase):
         self.assertEqual(_raw.dtype, _data.dtype)
 
     def test_export_to_file__32bit_uint(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         _raw = (np.random.random((11, 12, 13)) * 12573).astype(np.uint32)
         TiffIo.export_to_file(_fname, _raw)
         _data = TiffIo.import_from_file(_fname)
@@ -136,7 +147,7 @@ class TestTiffIo(unittest.TestCase):
         self.assertEqual(_raw.dtype, _data.dtype)
 
     def test_export_to_file__16bit_float(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         _raw = (np.random.random((11, 12, 13))).astype(np.float16)
         TiffIo.export_to_file(_fname, _raw)
         _data = TiffIo.import_from_file(_fname)
@@ -144,7 +155,7 @@ class TestTiffIo(unittest.TestCase):
         self.assertEqual(_raw.dtype, _data.dtype)
 
     def test_export_to_file__32bit_float(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         _raw = (np.random.random((11, 12, 13))).astype(np.float32)
         TiffIo.export_to_file(_fname, _raw)
         _data = TiffIo.import_from_file(_fname)
@@ -152,7 +163,7 @@ class TestTiffIo(unittest.TestCase):
         self.assertEqual(_raw.dtype, _data.dtype)
 
     def test_export_to_file__64bit_float(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         _raw = (np.random.random((11, 12, 13))).astype(np.float64)
         TiffIo.export_to_file(_fname, _raw)
         _data = TiffIo.import_from_file(_fname)
@@ -160,7 +171,7 @@ class TestTiffIo(unittest.TestCase):
         self.assertEqual(_data.dtype, np.float32)
 
     def test_export_to_file__1282bit_float(self):
-        _fname = os.path.join(self._path, self.get_fname())
+        _fname = self._path.joinpath(self.get_fname())
         _raw = (np.random.random((11, 12, 13))).astype(np.longfloat)
         TiffIo.export_to_file(_fname, _raw)
         _data = TiffIo.import_from_file(_fname)